Bug fixing according to tests

+++ src/main/kotlin/com/costular/sunkalc/MoonIllumination.kt
@@ -7,7 +7,7 @@ package com.costular.sunkalc
  * @property angle midpoint angle in radians of the illuminated limb of the moon reckoned eastward from the north point of the disk; the moon is waxing if the angle is negative, and waning if positive
  */
 data class MoonIllumination(
-    val fraction: Float,
-    val phase: Float,
+    val fraction: Double,
+    val phase: Double,
     val angle: Double
 )
